split resolvers and types
