+++ .eslintignore
@@ -1,0 +1,2 @@
+/graphql/generated/*
+/src/design/@generated/*
