+++ graphql/schema.graphql
@@ -6,6 +6,7 @@ type Activity {
   createdAt: DateTime!
   ghost: Ghost
   id: ID!
+  tags: [Tag]
   updatedAt: DateTime!
   user: User
 }
@@ -58,10 +59,12 @@ type Goal {
   participants: [Activity]
   personal: Boolean
   private: Boolean
-  project: [Project]
+  project: Project
   projectId: Int
   relatedTo: [Goal]
+  state: State
   stateId: String
+  tags: [Tag]
   title: String!
   updatedAt: DateTime!
 }
@@ -75,6 +78,7 @@ input GoalEstimate {
 type Mutation {
   createGoal(description: String!, estimate: GoalEstimate, key: Boolean, ownerId: String!, personal: Boolean, private: Boolean, projectId: Int!, stateId: String, title: String!, user: UserSession!): Goal
   createProject(description: String, flowId: String!, ownerId: String!, title: String!, user: UserSession!): Project
+  createTag(activityId: String!, color: String!, description: String, title: String!, user: UserSession!): Tag
   inviteUser(email: String!, user: UserSession!): Ghost
 }
 
@@ -87,13 +91,12 @@ type Project {
   id: Int!
   owner: Activity
   slug: String
+  tags: [Tag]
   title: String!
   updatedAt: DateTime!
 }
 
 type Query {
-  findGhost(query: String!, sortBy: SortOrder): [Ghost]
-  findUser(query: String!, sortBy: SortOrder): [User]
   findUserAnyKind(query: String!, sortBy: SortOrder): [UserAnyKind]
   flow(id: String!): Flow
   flowCompletion(query: String!, sortBy: SortOrder): [Flow]
@@ -118,6 +121,17 @@ type State {
   default: Boolean
   flows: [Flow]
   id: ID!
+  title: String!
+}
+
+type Tag {
+  activity: Activity
+  activityId: String
+  color: String!
+  description: String
+  goals: [Goal]
+  id: ID!
+  projects: [Project]
   title: String!
 }
